Move BNF Mapper bases to `bnf_mapper`
As a response to a pull request request.
