+++ web/modules/custom/bnf/src/Plugin/bnf_mapper/ParagraphGoVideoBundleAutomaticMapper.php
@@ -6,7 +6,7 @@ namespace Drupal\bnf\Plugin\bnf_mapper;
 
 use Drupal\bnf\Attribute\BnfMapper;
 use Drupal\bnf\GraphQL\Operations\GetNode\Node\Paragraphs\ParagraphGoVideoBundleAutomatic;
-use Drupal\bnf\Plugin\BnfMapperParagraphPluginBase;
+
 use Drupal\bnf\Plugin\Traits\EmbedVideoTrait;
 use Spawnia\Sailor\ObjectLike;
