+++ web/modules/custom/bnf/src/Plugin/bnf_mapper/ParagraphMediasMapper.php
@@ -6,7 +6,7 @@ namespace Drupal\bnf\Plugin\bnf_mapper;
 
 use Drupal\bnf\Attribute\BnfMapper;
 use Drupal\bnf\GraphQL\Operations\GetNode\Node\Paragraphs\ParagraphMedias;
-use Drupal\bnf\Plugin\BnfMapperParagraphPluginBase;
+
 use Drupal\bnf\Plugin\Traits\ImageTrait;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\File\FileSystemInterface;
